fix(P6): regenerate Notes verifier from current zkey + settle fallback for empty USDC
- Regenerate ConfidentialNotesVerifier.sol from notes_final.zkey (zkey was
  recompiled Jul 17 but verifier was not redeployed, causing InvalidProof
  reverts on createNote — TX1, TX2)
- Regenerate SpendVerifier.sol from spend_final.zkey (same mismatch)
- Fix settle endpoint: pre-check NoteSettlement USDC balance before PATH 1;
  fall through to PATH 2 (direct relayer transfer + MongoDB guard) when
  NoteSettlement has insufficient USDC (TX3 revert cause)
- Add DeployNotesVerifier.s.sol + deploy_notes_verifier.sh for the owner
  to deploy the new verifier + ConfidentialNotes on Base mainnet
- Add notes_post_deploy.py to auto-update server.py/confidential-notes.js/deployed_base.json

+++ contracts/src/SpendVerifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16Verifier {
     // Scalar field size
-    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax = 4046810554511096347223471314503653405319010211549719891067341276410419554538;
-    uint256 constant alphay = 20972216556230368226129142191999290211441534694053991095223717295099065119396;
-    uint256 constant betax1 = 1162525274628084786254208382708110717945079403327240881866527625099479139524;
-    uint256 constant betax2 = 17006775537206652401980775245876252329592576393762318603098098078744078304278;
-    uint256 constant betay1 = 728736684120689475367658410355586091946600404348752493917377386977690379436;
-    uint256 constant betay2 = 11247183439082070242085762446097360082332904929189577386647035313668925265922;
+    uint256 constant alphax  = 4046810554511096347223471314503653405319010211549719891067341276410419554538;
+    uint256 constant alphay  = 20972216556230368226129142191999290211441534694053991095223717295099065119396;
+    uint256 constant betax1  = 1162525274628084786254208382708110717945079403327240881866527625099479139524;
+    uint256 constant betax2  = 17006775537206652401980775245876252329592576393762318603098098078744078304278;
+    uint256 constant betay1  = 728736684120689475367658410355586091946600404348752493917377386977690379436;
+    uint256 constant betay2  = 11247183439082070242085762446097360082332904929189577386647035313668925265922;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,27 +42,24 @@ contract Groth16Verifier {
     uint256 constant deltay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
     uint256 constant deltay2 = 8495653923123431417604973247489272438418190587263600148770280649306958101930;
 
+    
     uint256 constant IC0x = 19990482892476878740126458503453962300775461854169600167181819997058074525496;
     uint256 constant IC0y = 9757298501629033459697729874116695658708355447314478280962806518794166633796;
-
+    
     uint256 constant IC1x = 16361379650382344942236778606186261389131590286619085272999386268280239524257;
     uint256 constant IC1y = 12791125120324735576718375507694040152955530868605706721344925848311473734887;
-
+    
     uint256 constant IC2x = 16622553612108139825555129119212595668262093364285399545608194975810391772565;
     uint256 constant IC2y = 515714054330325276873935415534916623977776394369627180930435662657653722291;
-
+    
+ 
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(
-        uint256[2] calldata _pA,
-        uint256[2][2] calldata _pB,
-        uint256[2] calldata _pC,
-        uint256[2] calldata _pubSignals
-    ) public view returns (bool) {
+    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[2] calldata _pubSignals) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -70,7 +67,7 @@ contract Groth16Verifier {
                     return(0, 0x20)
                 }
             }
-
+            
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -105,10 +102,11 @@ contract Groth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-
+                
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-
+                
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
+                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -134,6 +132,7 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
+
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -150,6 +149,7 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
+
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -159,16 +159,17 @@ contract Groth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-
+            
             checkField(calldataload(add(_pubSignals, 0)))
-
+            
             checkField(calldataload(add(_pubSignals, 32)))
+            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-            return(0, 0x20)
-        }
-    }
-}
+             return(0, 0x20)
+         }
+     }
+ }
